% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{komma}
\alias{komma}
\title{Komma}
\usage{
komma(...)
}
\arguments{
\item{...}{passed on to scales::comma_format}
}
\value{
a function to use in label argument in scales in ggplot2
}
\description{
Komma
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(drat)) +
geom_histogram() +
scale_x_continuous(labels = komma()) + theme_minimal()
}
