% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pederlib2-package.R
\docType{package}
\name{pederlib2-package}
\alias{pederlib2}
\alias{pederlib2-package}
\title{pederlib2: Personal package take two}
\description{
Personal package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedersebastian.github.io/pederlib2/}
}

}
\author{
\strong{Maintainer}: Peder Sebastian Martinsen \email{peder.s.martinsen@gmail.com}

}
\keyword{internal}
