% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pederlib2-package.R
\docType{package}
\name{pederlib2-package}
\alias{pederlib2}
\alias{pederlib2-package}
\title{pederlib2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
