% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{startup}
\alias{startup}
\title{Start up}
\usage{
startup(type, ...)
}
\arguments{
\item{type}{with type}

\item{...}{add more packs}
}
\value{
list of attached packages
}
\description{
Start up
}
\examples{

startup()
}
