% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{startup}
\alias{startup}
\title{Start up}
\usage{
startup(..., quiet = FALSE)
}
\arguments{
\item{...}{packs to add}

\item{quiet}{description}
}
\value{
list of attached packages
}
\description{
type = 1 - tidyverse + readxl
type = 2 - tidymodels core pkg TODO
type = 3 - tidymodel core + ext TODO
type = 4 - web TODO
}
\examples{

startup()
}
