% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pedr.R
\name{theme_pedr}
\alias{theme_pedr}
\title{Theme_pedr}
\usage{
theme_pedr(
  base_size = 11,
  strip_text_size = 12,
  strip_text_margin = 5,
  subtitle_size = 13,
  subtitle_margin = 10,
  plot_title_size = 16,
  plot_title_margin = 10,
  strip_color = "gray90",
  font_family = "BentonSans Regular",
  ...
)
}
\arguments{
\item{base_size}{x}

\item{strip_text_size}{x}

\item{strip_text_margin}{x}

\item{subtitle_size}{x}

\item{subtitle_margin}{x}

\item{plot_title_size}{x}

\item{plot_title_margin}{x}

\item{strip_color}{x}

\item{font_family}{x}

\item{...}{passed on to theme_minimal}
}
\value{
theme
}
\description{
Theme_pedr
}
\examples{

library(ggplot2)
ggplot(mtcars, aes(mpg, disp)) +
geom_point() +
facet_wrap(vars(vs)) +
theme_pedr(font_family = "Times New Roman")

}
